chore: address automation targeting review cleanup
