+++ src/renderer/src/lib/launch-agent-background-session.ts
@@ -1,18 +1,20 @@
-import { toast } from 'sonner'
 import { useAppStore } from '@/store'
 import { buildAgentStartupPlan, type AgentStartupPlan } from '@/lib/tui-agent-startup'
+import type {
+  LaunchAgentBackgroundSessionArgs,
+  LaunchAgentBackgroundSessionResult
+} from '@/lib/agent-background-session-contract'
 import { getAgentLaunchPlatformForRepo } from '@/lib/agent-launch-platform'
 import { CLIENT_PLATFORM } from '@/lib/new-workspace'
-import { track, tuiAgentToAgentKind } from '@/lib/telemetry'
+import { tuiAgentToAgentKind } from '@/lib/telemetry'
 import { pasteDraftWhenAgentReady } from '@/lib/agent-paste-draft'
+import { showAutomationPromptNotSentToast } from '@/lib/agent-background-session-timeout-toast'
 import { getLocalProjectExecutionRuntimeContext } from '@/lib/local-preflight-context'
 import {
   resolveTuiAgentLaunchArgs,
   resolveTuiAgentLaunchEnv
 } from '../../../shared/tui-agent-launch-defaults'
 import { TUI_AGENT_CONFIG } from '../../../shared/tui-agent-config'
-import type { TuiAgent } from '../../../shared/types'
-import type { LaunchSource } from '../../../shared/telemetry-events'
 import { makePaneKey } from '../../../shared/stable-pane-id'
 import {
   registerEagerPtyBuffer,
@@ -30,28 +32,10 @@ import {
   toRemoteRuntimePtyId
 } from '@/runtime/runtime-terminal-stream'
 import { createAgentStatusOscProcessor } from '../../../shared/agent-status-osc'
-import type { ParsedAgentStatusPayload } from '../../../shared/agent-status-types'
 import type { RuntimeTerminalCreate } from '../../../shared/runtime-types'
-import { translate } from '@/i18n/i18n'
 import { createSshBackgroundStartupDelivery } from '@/lib/ssh-background-startup-delivery'
 import { shouldUseShellReadyStartupDelivery } from '../../../shared/codex-startup-delivery'
 
-export type LaunchAgentBackgroundSessionArgs = {
-  agent: TuiAgent
-  worktreeId: string
-  prompt?: string
-  launchSource?: LaunchSource
-  title?: string
-  onData?: (chunk: string) => void
-  onExit?: (ptyId: string, code: number) => void
-  onAgentStatus?: (payload: ParsedAgentStatusPayload) => void
-}
-export type LaunchAgentBackgroundSessionResult = {
-  tabId: string
-  paneKey: string
-  ptyId: string
-  startupPlan: AgentStartupPlan
-}
 export async function launchAgentBackgroundSession(
   args: LaunchAgentBackgroundSessionArgs
 ): Promise<LaunchAgentBackgroundSessionResult | null> {
@@ -305,18 +289,7 @@ export async function launchAgentBackgroundSession(
       content: pasteDraftAfterLaunch,
       agent,
       submit: true,
-      onTimeout: () => {
-        toast.message(
-          translate(
-            'auto.lib.launch.agent.background.session.4ca0651d56',
-            "Your automation prompt wasn't sent — open the workspace and paste it."
-          )
-        )
-        track('agent_error', {
-          error_class: 'paste_readiness_timeout',
-          agent_kind: tuiAgentToAgentKind(agent)
-        })
-      }
+      onTimeout: () => showAutomationPromptNotSentToast(agent)
     })
   }
